backend/requirements.txt
Flask
Flask-Cors
spacy
dateparser
requests
python-dotenv
email-validator
pytest
